fix(android): StrokeDescription 时间参数改 Long，修 Kotlin 编译错误
GestureDescription.StrokeDescription 构造签名 (Path, startTime: Long,
duration: Long)，Int 型 duration 需显式 toLong()；startTime 字面量
补 L。此错误此前被 C++ 层编译失败遮蔽，NDK 修通后首次暴露。

+++ apps/android/app/src/main/java/com/wingman/agent/WingmanAccessibilityService.kt
@@ -45,7 +45,7 @@ class WingmanAccessibilityService : AccessibilityService() {
                 lineTo(x2.toFloat(), y2.toFloat())
             }
         }
-        val stroke = GestureDescription.StrokeDescription(path, 0, duration)
+        val stroke = GestureDescription.StrokeDescription(path, 0L, duration.toLong())
         val gesture = GestureDescription.Builder().addStroke(stroke).build()
         val dispatched = dispatchGesture(gesture, object : GestureResultCallback() {
             override fun onCompleted(gestureDescription: GestureDescription?) {
